requirements.txt
# Core ML Libraries
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
peft>=0.7.0

# Training Utilities
trl>=0.7.0
bitsandbytes>=0.41.0
flash-attn>=2.3.0
deepspeed>=0.12.0

# HuggingFace Integration
huggingface-hub>=0.20.0
tokenizers>=0.15.0
safetensors>=0.4.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0
sentencepiece>=0.1.99

# Monitoring & Logging
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Utilities
pyyaml>=6.0.0
omegaconf>=2.3.0
rich>=13.0.0
typer>=0.9.0

# Evaluation
evaluate>=0.4.0
nltk>=3.8.0
rouge-score>=0.1.2
